+++ samples/Demo.StaThreads.Test/ScreenObjects/MessageBoxScreen.cs
@@ -1,0 +1,29 @@
+﻿using FlaUI.Core.AutomationElements;
+using TestBaseClasses;
+
+namespace Demo.StaThreads.ScreenObjects;
+
+public class MessageBoxScreen : Screen
+{
+    public MessageBoxScreen(Window window)
+        : base(window)
+    {
+    }
+
+    private Label MessageLabel => ElementByAutomationId<Label>("65535");
+
+    public string Caption => Window.Title;
+
+    public string Message => MessageLabel.Text;
+
+    public bool IsIconVisible => ElementWithAutomationIdExists("20");
+        
+    public bool IsOKButtonVisible => ElementWithTextExists(OK);
+
+    public bool IsCancelButtonVisible => ElementWithTextExists(Cancel);
+
+    public void ClickOK()
+    {
+        DefaultOKButton.Click();
+    }
+}